+++ ft_memccpy.c
@@ -6,7 +6,7 @@
 /*   By: alboumed <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/10/12 11:08:44 by alboumed          #+#    #+#             */
-/*   Updated: 2019/10/22 20:21:52 by alboumed         ###   ########.fr       */
+/*   Updated: 2019/10/22 20:41:26 by alboumed         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	*ft_memccpy(void *dest, const void *src, int c, size_t n)
 {
-	int i;
+	size_t i;
 
 	i = 0;
 	while(((char*)src)[i] != c && i < n)
