+++ ft_memchr.c
@@ -6,7 +6,7 @@
 /*   By: alboumed <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/10/13 14:42:48 by alboumed          #+#    #+#             */
-/*   Updated: 2019/10/22 20:22:06 by alboumed         ###   ########.fr       */
+/*   Updated: 2019/10/22 20:41:51 by alboumed         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	*ft_memchr(const void *s, int c, size_t n)
 {
-	int i;
+	size_t i;
 
 	i = 0;
 	while (i < n && (unsigned char)c != ((unsigned char*)s)[i])
