+++ ft_memcmp.c
@@ -6,7 +6,7 @@
 /*   By: alboumed <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/10/13 15:09:45 by alboumed          #+#    #+#             */
-/*   Updated: 2019/10/22 20:22:22 by alboumed         ###   ########.fr       */
+/*   Updated: 2019/10/22 20:42:28 by alboumed         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 int		ft_memcmp(const void *s1, const void *s2, size_t n)
 {
-	int i;
+	size_t i;
 
 	i = 0;
 	while (((unsigned char*)s1)[i] == ((unsigned char*)s2)[i] && i < n)
