+++ ft_memset.c
@@ -6,7 +6,7 @@
 /*   By: alboumed <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/10/11 10:49:37 by alboumed          #+#    #+#             */
-/*   Updated: 2019/10/22 20:23:17 by alboumed         ###   ########.fr       */
+/*   Updated: 2019/10/23 12:33:06 by alboumed         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	*ft_memset(void *b, int c, size_t len)
 {
-	int i;
+	size_t i;
 
 	i = 0;
 	while (i < len)
